ItemView,ModelData and ItemRow are edited. JSONfiles are edited. load func is redesigned. Next, I need to make func to load wepons data and display these list.
