+++ m8sp104/ViewModel/eventDates.json
@@ -2,13 +2,13 @@
     {
         "id": 1,
         "name": "原初の闇",
-        "startDate": [2021,1,5,23,30],
-        "endDate": [2021,1,6,0,0]
+        "startDate": "2021-01-05 23:30:00",
+        "endDate": "2021-01-06 00:00:00"
     },
     {
         "id": 2,
         "name": "メリクリ2020",
-        "startDate": [2021,1,6,4,0],
-        "endDate": [2021,1,6,5,30]
+        "startDate": "2021-01-06 04:30:00",
+        "endDate": "2021-01-06 05:30:00"
     }
 ]
