--- m8sp104/ViewModel/sample.json
@@ -1,6 +1,0 @@
-
-[
-    {
-        "name": "sword"
-    }
-]
